Code Review improvements // embedded all classes, having Log as their only purpose, to the Log class to use as namespace
