--- Alicerce/Sources/Logging/Formatters/LogItemLevelDefaultFormatter.swift
@@ -1,11 +1,0 @@
-//
-//  LogItemLevelDefaultFormatter.swift
-//  Alicerce
-//
-//  Created by Meik Schutz on 07/04/2017.
-//  Copyright © 2017 Mindera. All rights reserved.
-//
-
-public final class LogItemLevelDefaultFormatter: LogItemLevelFormatter {
-
-}
