+++ Alicerce/Sources/Logging/Log+ItemLevelDefaultFormatter.swift
@@ -1,0 +1,15 @@
+//
+//  Log+ItemLevelDefaultFormatter.swift
+//  Alicerce
+//
+//  Created by Meik Schutz on 07/04/2017.
+//  Copyright © 2017 Mindera. All rights reserved.
+//
+
+public extension Log {
+
+    public final class LogItemLevelDefaultFormatter: LogItemLevelFormatter {
+        
+    }
+}
+
