+++ Alicerce/Sources/Logging/Protocols/LogItemLevelFormatter.swift
